+++ balashov.dima/P3/matrixProcessing.hpp
@@ -7,8 +7,8 @@
 
 namespace balashov
 {
-  double minSumSideDiagonal(int matrix[], size_t column, size_t row);
-  double diagonalCalculation(int matrix[], int columnNow, size_t columnMax,
+  int minSumSideDiagonal(int matrix[], size_t column, size_t row);
+  int diagonalCalculation(int matrix[], int columnNow, size_t columnMax,
     int rowNow, size_t rowMax);
 }
